+++ lib/datacite/mapping/identifier.rb
@@ -2,7 +2,7 @@ require 'xml/mapping'
 
 module Datacite
   module Mapping
-    DOI_PATTERN = %r{10\..+/.+}
+    DOI_PATTERN = %r{\A10\.\S+/\S+\z}
 
     # The persistent identifier that identifies the resource.
     #
@@ -24,9 +24,10 @@ module Datacite
       end
 
       def value=(v)
-        fail ArgumentError, 'Identifier must have a non-nil value' unless v
-        fail ArgumentError, "Identifier value '#{v}' is not a valid DOI" unless v.match(DOI_PATTERN)
-        @value = v
+        new_value = v && v.strip
+        fail ArgumentError, 'Identifier must have a non-nil value' unless new_value
+        fail ArgumentError, "Identifier value '#{new_value}' is not a valid DOI" unless new_value.match(DOI_PATTERN)
+        @value = new_value
       end
 
       # Sets the identifier type. Should only be called by the XML mapping engine.
@@ -50,18 +51,30 @@ module Datacite
         Identifier.new(value: match[0])
       end
 
-      def self.extract_and_set_value(obj, xml)
-        obj.value = xml.text
-      end
-      private_class_method :extract_and_set_value
-
+      text_node :value, 'text()'
       text_node :identifier_type, '@identifierType'
-
-      # The default reader throws an opaque XPath error if there's no element text,
-      # so we extract it manually and let the accessor raise an error instead
-      text_node :value, 'text()', reader: method(:extract_and_set_value)
 
       fallback_mapping :datacite_3, :_default
     end
+
+    # Custom node to warn (but not blow up) if we read an XML `<resource/>` that's
+    # missing its `<identifier/>`.
+    class IdentifierNode < XML::Mapping::ObjectNode
+      def xml_to_obj(_obj, xml)
+        element = element_from(xml)
+        return unless element
+        return super if element.text
+        warn "Missing identifier value in #{element}; add a valid Identifier to the resulting Resource before saving"
+      end
+
+      private
+
+      def element_from(xml)
+        @path.first(xml)
+      rescue XML::XXPathError => e
+        warn "Identifier element #{@attrname} not found in XML #{xml}: #{e}; add a valid Identifier to the resulting Resource before saving"
+      end
+    end
+    XML::Mapping.add_node_class IdentifierNode
   end
 end